refactor(abs): Condense `abs_min` and `abs_max`
