+++ maths/abs.py
@@ -13,6 +13,61 @@ def abs_val(num: float) -> float:
     0
     """
     return -num if num < 0 else num
+    
+
+def abs_min(x: list[int]) -> int:
+    """
+    >>> abs_min([0,5,1,11])
+    0
+    >>> abs_min([3,-10,-2])
+    -2
+    >>> abs_min([])
+    Traceback (most recent call last):
+        ...
+    ValueError: abs_min() arg is an empty sequence
+    """
+    if len(x) == 0:
+        raise ValueError("abs_min() arg is an empty sequence")
+    j = x[0]
+    for i in x:
+        if abs_val(i) < abs_val(j):
+            j = i
+    return j
+
+def abs_max(x: list[int]) -> int:
+    """
+    >>> abs_max([0,5,1,11])
+    11
+    >>> abs_max([3,-10,-2])
+    -10
+    >>> abs_max([])
+    Traceback (most recent call last):
+        ...
+    ValueError: abs_max() arg is an empty sequence
+    """
+    if len(x) == 0:
+        raise ValueError("abs_max() arg is an empty sequence")
+    j = x[0]
+    for i in x:
+        if abs(i) > abs(j):
+            j = i
+    return j
+
+
+def abs_max_sort(x: list[int]) -> int:
+    """
+    >>> abs_max_sort([0,5,1,11])
+    11
+    >>> abs_max_sort([3,-10,-2])
+    -10
+    >>> abs_max_sort([])
+    Traceback (most recent call last):
+        ...
+    ValueError: abs_max_sort() arg is an empty sequence
+    """
+    if len(x) == 0:
+        raise ValueError("abs_max_sort() arg is an empty sequence")
+    return sorted(x, key=abs)[-1]
 
 
 def test_abs_val():
@@ -23,6 +78,16 @@ def test_abs_val():
     assert 34 == abs_val(34)
     assert 100000000000 == abs_val(-100000000000)
 
+    a = [-3, -1, 2, -11]
+    assert abs_max(a) == -11
+    assert abs_max_sort(a) == -11
+    assert abs_min(a) == -1
+
 
 if __name__ == "__main__":
+    import doctest
+
+    doctest.testmod()
+
+    test_abs_val()
     print(abs_val(-34))  # --> 34
